apps/amp-support-bot/pyproject.toml
[tool.setuptools]
packages = ["app"]

[project]
name = "amp-support-bot"
version = "1.0.0"
description = "Grounded RAG support bot with FastAPI, Ollama, and Chroma."
authors = [{ name = "Amplivo CSRnow", email = "support@example.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "pydantic==2.9.2",
    "httpx==0.27.2",
    "chromadb==0.5.11",
    "markdown-it-py==3.0.0",
    "python-dotenv==1.0.1",
    "python-docx==1.1.0",
    "PyMuPDF==1.24.9", 
    
    
]
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
